requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-migrate==4.0.7
flask-login==0.6.3
flask-bcrypt==1.0.1
flask-wtf==1.2.1
psycopg2-binary==2.9.9
cloudinary==1.40.0
gunicorn==22.0.0
python-dotenv==1.0.1
